# routes/auth.js
const router = require("express").Router();
const User = require("../model/User");
const bcrypt = require("bcryptjs");
const jwt = require("jsonwebtoken");

const auth = require('../middleware/auth')
const { check, validationResult } = require("express-validator");

router.get('/', auth, async (req, res) => {

  try {
    
    const user = await User.findById(req.user.id).select('-password')
    
    if(user == null){
      res.status(404).json({ msg: 'User Not Longer Exist'})
    }
    
    res.json(user)

  } catch (error) {
    console.log(error.message)
    res.status(500).send('Internal Server Error')
  }

})

router.post(
  '/login',
  [
    check('email', 'Please include a valid email').isEmail(),
    check('password', 'Password is required').exists()
  ],
  async (req, res) => {
    const errors = validationResult(req);
    if (!errors.isEmpty()) {
      return res.status(400).json({ errors: errors.array() });
    }

    const { email, password } = req.body;

    try {
      let user = await User.findOne({ email });

      if (!user) {
        return res
          .status(400)
          .json({ errors: [{ msg: 'Invalid Credentials' }] });
      }

      const isMatch = await bcrypt.compare(password, user.password);

      if (!isMatch) {
        return res
          .status(400)
          .json({ errors: [{ msg: 'Invalid Credentials' }] });
      }

      const payload = {
        user: {
          id: user.id
        }
      };

      const secretKey = 'secret'

      jwt.sign(
        payload,
        secretKey,
        { expiresIn: '5 days' },
        (err, token) => {
          if (err) throw err;
          res.status(200).json({ token });
        }
      );
    } catch (err) {
      console.error(err.message);
      res.status(500).send('Internal Server error');
    }
  }
);

module.exports = router;